Split Sun context owners
