+++ service-worker.js
@@ -506,7 +506,10 @@ const APP_SHELL = [
   '/js/sun-body-silhouette.js',
   '/js/sun-ai-analysis.js',
   '/js/sun-context.js',
+  '/js/sun-context-environment.js',
   '/js/sun-context-hooks.js',
+  '/js/sun-context-runtime.js',
+  '/js/sun-context-session-tools.js',
   '/js/sun-correlations.js',
   '/js/sun-defaults.js',
   '/js/sun-defaults-runtime.js',
