+++ tests/test-quality-guardrails.js
@@ -250,6 +250,9 @@ const domainUiCheckJsModules = [
   'js/pdf-import-review.js',
   'js/provider-panels.js',
   'js/sun-context.js',
+  'js/sun-context-environment.js',
+  'js/sun-context-runtime.js',
+  'js/sun-context-session-tools.js',
   'js/sun-defaults.js',
   'js/sun-defaults-runtime.js',
   'js/sun-location.js',
