+++ tsconfig.json
@@ -286,6 +286,9 @@
     "js/sun-body-silhouette.js",
     "js/sun-channel-metrics.js",
     "js/sun-context.js",
+    "js/sun-context-environment.js",
+    "js/sun-context-runtime.js",
+    "js/sun-context-session-tools.js",
     "js/sun-correlations.js",
     "js/sun-defaults.js",
     "js/sun-defaults-runtime.js",
